% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetterMethods.R
\docType{methods}
\name{getter}
\alias{getter}
\alias{$,MEITests-method}
\title{Applies method $ to different classes}
\usage{
\S4method{$}{MEITests}(x, name)
}
\arguments{
\item{x}{Object.}

\item{name}{Name of slot.}
}
\description{
Enables slots of objects in this package to be referenced easily.
}
\keyword{internal}
