% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-MEITests-class.R
\docType{class}
\name{MEITests-class}
\alias{MEITests-class}
\title{S4 class for MEI tests results}
\description{
S4 class for MEI tests results
}
\section{Slots}{

\describe{
\item{\code{Zvalue_major}}{Numeric}

\item{\code{Pvalue_major}}{Numeric}

\item{\code{Zvalue_normal}}{Numeric}

\item{\code{Pvalue_normal}}{Numeric}

\item{\code{Pvalue_combine}}{Numeric}

\item{\code{SNPsNum}}{Numeric}

\item{\code{sel.pthr}}{Numeric}
}}

